\name{SPORTSCausal-package}
\alias{SPORTSCausal-package}
\alias{SPORTSCausal}
\docType{package}
\title{
\packageTitle{SPORTSCausal}
}
\description{
\packageDescription{SPORTSCausal}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SPORTSCausal}
\packageIndices{SPORTSCausal}

}
\author{
\packageAuthor{SPORTSCausal}

Maintainer: \packageMaintainer{SPORTSCausal}
}
\references{
Brodersen et al. \emph{Inferring causal impact using Bayesian structural time-series models}. Annals of Applied Statistics, 2015
}
\keyword{ package }
\seealso{
\url{http://google.github.io/CausalImpact/CausalImpact.html}
}
\examples{
## For more detail of the package, try ?sportscausal and ?ad_cost
}
